99.99% just specifc color/size

+++ src/Components/citationCreator.js
@@ -5,43 +5,50 @@ const CitationCreationTool = (props) => {
   const [citationInput, setCitationInput] = useState("");
   const [refernceInput, setRefernceInput] = useState("");
   const [fileInput, setFileInput] = useState();
-  const [snanButtonText, setSnanButtonText] = useState("Scan Selected Image");
-  function createContainerFunction() {
-    const accountObject = props.accounts.filter((account) => {
-      return account.username === props.loggedInAccountUsername;
-    })[0];
-    const date = new Date();
+  const [snanButtonText, setSnanButtonText] = useState("Scan Image");
+  function createCitationFunction() {
+    if (citationInput !== "" && refernceInput !== "") {
+      const accountObject = props.accounts.filter((account) => {
+        return account.username === props.loggedInAccountUsername;
+      })[0];
+      const date = new Date();
 
-    let day = date.getDate();
-    let month = date.getMonth() + 1;
-    let year = date.getFullYear();
+      let day = date.getDate();
+      let month = date.getMonth() + 1;
+      let year = date.getFullYear();
 
-    let currentDate = `${day}/${month}/${year}`;
+      let currentDate = `${day}/${month}/${year}`;
 
-    accountObject.quotes.push({
-      citation: citationInput,
-      reference: refernceInput,
-      container: props.insideContainer,
-      favourite: false,
-      date: currentDate,
-    });
-    props.setAccounts((account) => {
-      return account.username != props.loggedInAccountUsername;
-    });
-    props.setAccounts([...props.accounts, accountObject]);
-    setCitationInput("");
-    setRefernceInput("");
+      accountObject.quotes.push({
+        citation: citationInput,
+        reference: refernceInput,
+        container: props.insideContainer,
+        favourite: false,
+        date: currentDate,
+      });
+      props.setAccounts((account) => {
+        return account.username != props.loggedInAccountUsername;
+      });
+      props.setAccounts([...props.accounts, accountObject]);
+      setCitationInput("");
+      setRefernceInput("");
+    } else {
+      setCitationInput("");
+      setRefernceInput("");
+    }
   }
 
   async function scanImage() {
-    const beforeCitation = citationInput;
-    setSnanButtonText("Loading");
-    const worker = await createWorker("eng");
-    const ret = await worker.recognize(fileInput);
-    setCitationInput(ret.data.text);
-    setSnanButtonText("Scan Item");
-    setFileInput(undefined);
-    await worker.terminate();
+    if (fileInput !== undefined) {
+      const beforeCitation = citationInput;
+      setSnanButtonText("Loading");
+      const worker = await createWorker("eng");
+      const ret = await worker.recognize(fileInput);
+      setCitationInput(ret.data.text);
+      setSnanButtonText("Scan Image");
+      setFileInput(undefined);
+      await worker.terminate();
+    }
   }
   return (
     <div className="citationCreatorContainer">
@@ -66,7 +73,10 @@ const CitationCreationTool = (props) => {
           <label
             htmlFor="imageInput"
             className="imageInputLabel"
-            style={{ display: fileInput == undefined ? null : "none" }}
+            style={{
+              border:
+                fileInput == undefined ? "1px dashed black" : "1px solid black",
+            }}
           >
             <p className="imageInputTitle">
               {fileInput == undefined
@@ -75,20 +85,18 @@ const CitationCreationTool = (props) => {
             </p>
           </label>{" "}
           <div
-            style={{ display: fileInput == undefined ? "none" : null }}
             className="scanImageButton"
             onClick={() => {
               scanImage();
             }}
           >
-            {snanButtonText}
+            <h3 className="scanButtonTitle">{snanButtonText}</h3>
           </div>{" "}
           <div
             className="createCitationButton"
-            onClick={createContainerFunction}
+            onClick={createCitationFunction}
           >
-            <h5 className="createCitationButtonTitle1">Create </h5>{" "}
-            <h5 className="createCitationButtonTitle2"> Citation</h5>
+            <h5 className="createCitationButtonTitle1">Create Citation</h5>{" "}
           </div>
         </div>
       </div>
